refactor: Use Float64 in psi
Using a Float16 in the calculation of the potential seems to
occasionally cause problems. Use Float64 everywhere for now.

+++ src/cylindrical.jl
@@ -35,7 +35,7 @@ maximum(p.R)
 struct CylindricalProfile
     R::Array{Float64,2}
     z::Array{Float64,2}
-    psi::Array{Complex{Float16},3}
+    psi::Array{Complex{Float64},3}
 end
 
 function CylindricalProfile(
